User privilage and ordermanagement

+++ src/main/java/com/example/thisaraprinters/repository/EmployeeRepo.java
@@ -3,7 +3,6 @@ package com.example.thisaraprinters.repository;
 import com.example.thisaraprinters.model.EmployeeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepo extends JpaRepository<EmployeeModel, Long> {
+public interface EmployeeRepo extends JpaRepository<EmployeeModel, Integer> {
     boolean existsByNic(String nic);
-
 }
